Move publish and next_args into own modules
